+++ example/js/app/components/search/search-result.jsx
@@ -1,14 +1,15 @@
 import React, { Component } from 'react';
-import { Container, Grid } from '@/rext';
+import { Container, Grid, Dropdown } from '@/rext';
 import CardStore from '#/stores/card';
+import CardTypeStore from '#/stores/card-type';
 
-export default class extends Component {
+export default class SearchResult extends Component {
   render() {
     return <Container>
       <Grid store={CardStore}>
         <div text="ID" dataIndex="Id" className="text-center" style={{width:150}} />
-        <div text="Name" dataIndex="Name" className="text-center" style={{width:250}} />
-        <div text="Type" dataIndex="Type" className="text-center" style={{width:100}} />
+        <div text="Name" dataIndex="Name" className="text-center" style={{width:250}} editable />
+        <div text="Type" dataIndex="Type" className="text-center" style={{width:100}} editable={{type:'dropdown',store:CardTypeStore}} />
         <div text="Introduction" dataIndex="Introduction" className="text-center" style={{width:1000}} />
         <div text="STR" dataIndex="STR" className="text-center" style={{width:50}} />
         <div text="AGI" dataIndex="AGI" className="text-center" style={{width:50}} />