+++ example/js/app/stores/card-type.js
@@ -2,9 +2,5 @@ import { Store } from '@/rext'
 
 export default Store({
   storeId: 'CardTypeStore',
-  data: [
-    { id: null, code: 'MELEE', name: 'Melee' },
-    { id: null, code: 'SHOOT', name: 'Shoot' },
-    { id: null, code: 'MAGIC', name: 'Magic' },
-  ]
+  data: ['Melee', 'Shoot', 'Magic']
 });