+++ src/components/container.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-export default class extends Component {
+export default class Container extends Component {
   render() {
     const { className = '', hbox, children, ...others } = this.props;
     return <section className={`d-flex flex-${hbox ? 'row' : 'column'} ${className}`} {...others}>{children}</section>;