+++ src/core/string.js
@@ -9,6 +9,10 @@ class String {
     }
     return pairs.join(sep);
   }
+
+  capitalize(string) {
+    return string.charAt(0).toUpperCase() + string.slice(1);
+  }
 }
 
 export default new String;