+++ src/rext.js
@@ -9,6 +9,7 @@ class Rext {
   constructor() {
     this.extend = Ext.extend;
     this.ajax = (settings) => Ajax.request(settings);
+    this.generateSetter = Ext.generateSetter;
   }
 
   async launch(viewport) {